stib-waittime-source/requirements.txt
-e git+https://github.com/quixio/quix-streams.git@quent/sources#egg=quixstreams
python-dotenv
requests
